fix(kibbeh): fix #2371
